Added comma and semicolon delimiters to InputStreamMinesweeperPlayer and moved to player package
